resources/test/tests/functional/assert-array-equals.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(() => {
  assert_array_equals([], []);
}, "empty and equal");
test(() => {
  assert_array_equals([1], [1]);
}, "non-empty and equal");
test(() => {
  assert_array_equals([], [1]);
}, "length differs");
test(() => {
  assert_array_equals([1], [,]);
}, "property is present");
test(() => {
  assert_array_equals([,], [1]);
}, "property is missing");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20], ["x",1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20]);
}, "property 0 differs");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20], [0,1,2,3,4,"x",6,7,8,9,10,11,12,13,14,15,16,17,18,19,20]);
}, "property 5 differs");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21], [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]);
}, "lengths differ and input array beyond display limit");
test(() => {
  assert_array_equals([1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21], [0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]);
}, "lengths differ and expected array beyond display limit");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21], [,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]);
}, "property 0 is present and arrays are beyond display limit");
test(() => {
  assert_array_equals([,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21], [0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]);
}, "property 0 is missing and arrays are beyond display limit");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21], [0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,,19,20,21]);
}, "property 18 is present and arrays are beyond display limit");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,,19,20,21], [0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]);
}, "property 18 is missing and arrays are beyond display limit");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21], ["x",1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]);
}, "property 0 differs and arrays are beyond display limit");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21], [0,1,2,3,4,"x",6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]);
}, "property 5 differs and arrays are beyond display limit");
test(() => {
  assert_array_equals([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26], [0,1,2,3,4,"x",6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26]);
}, "property 5 differs and arrays are beyond display limit on both sides");

</script>
<script type="text/json" id="expected">
TODO I will clean this up once this is generally okay

{
  "summarized_status": {
    "status_string": "OK",
    "message": null
  },
  "summarized_tests": [
    {
      "status_string": "PASS",
      "name": "Cleanup methods are invoked exactly once and in the expected sequence.",
      "properties": {},
      "message": null
    },
    {
      "status_string": "PASS",
      "name": "Cleanup methods are invoked following the completion of asynchronous tests",
      "properties": {},
      "message": null
    },
    {
      "status_string": "PASS",
      "name": "probe asynchronous",
      "properties": {},
      "message": null
    },
    {
      "status_string": "PASS",
      "name": "probe synchronous",
      "properties": {},
      "message": null
    }
  ],
  "type": "complete"
}
</script>
